OD-0000 Fix compilation errors
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@49307 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ editor-api/src/main/groovy/ish/oncourse/willow/editor/rest/WebNodeTypeToTheme.groovy
@@ -18,7 +18,7 @@ class WebNodeTypeToTheme {
     Theme getTheme() {
         return new Theme().with { theme ->
             theme.title = webNodeType.name
-            theme.id = webNodeType.id
+            theme.id = webNodeType.id.doubleValue()
 //            theme.layoutId = webNodeType.webSiteLayout
             theme
         }
